Address PR #229 review: Phases 2-4 structural fixes
SessionConnector enum extraction:
- Move IlcaNetSessionState to externals/INL1/IlcaNetSessionState.h
- Move IlcaNetSessionNetworkType to externals/INL1/IlcaNetSessionNetworkType.h
- Move SessionEventType to externals/Dpr/NetworkUtils/SessionEventType.h
- Move SessionEventData to externals/Dpr/NetworkUtils/SessionEventData.h
- SessionConnector.h includes new files, uses 'using' for compatibility

OpcController fixes:
- Fix Vector2 field to use ::Fields instead of ::Object for proper
  ARM64 register passing (s0/s1)
- Extract UgOpcController to separate externals/UgOpcController.h

Merge GetEscapeMode hooks:
- Move team-up escape check from team_up.cpp into
  battle_escape_flag.cpp's MainModule_GetEscapeMode hook
- Eliminates hook conflict at 0x2034BB0 (both files hooked same addr)

File organization:
- Move overworld_mp_interact.cpp from commands/ to features/
  (it's interaction logic, not a script command)
