+++ src/mod/externals/Dpr/NetworkUtils/SessionEventType.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <cstdint>
+
+namespace Dpr::NetworkUtils {
+    enum class SessionEventType : int32_t {
+        Join = 0,
+        Leave = 1,
+    };
+}
