+++ src/mod/externals/INL1/IlcaNetSessionNetworkType.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <cstdint>
+
+namespace INL1 {
+    enum class IlcaNetSessionNetworkType : int32_t {
+        None = 0,
+        Lan = 1,
+        Internet = 2,
+    };
+}
